Did some more refactoring
