+++ src/js/libraryDeliveryPigeon.js
@@ -1,5 +1,5 @@
 //this module is used for most of the work done on the library.html page
-import { grabLocalStorage, feedLocalStorage } from "./flyInNutsAndBolts.js";
+import { grabLocalStorage, feedLocalStorage } from "./utils.js";
 import ExternalServices from "./externalServices.js";
 
 //create variable for ElectonicPigeon Module
