+++ src/js/main.js
@@ -1,5 +1,5 @@
 import ExternalServices from "./externalServices.js";
-import { flyInHeaderFooter, feedLocalStorage } from "./flyInNutsAndBolts.js";
+import { flyInHeaderFooter, feedLocalStorage } from "./utils.js";
 
 //put in the header & footer
 flyInHeaderFooter();
