+++ src/partials/header.html
@@ -1,4 +1,4 @@
-<div class="logo">
+<div>
   <nav class="navbar">
     <!-- LOGO -->
     <div class="logo">
@@ -26,10 +26,10 @@
           <a href="/">My Library</a>
           <!-- DROPDOWN MENU -->
           <ul class="dropdown">
-            <li><a href="library.html">My Library</a></li>
             <li><a href="/">Books to Read </a></li>
             <li><a href="/">Reading Now</a></li>
             <li><a href="/">Finished Books</a></li>
+            <li><a href="library.html">My Library</a></li>
           </ul>
         </li>
         <li><a href="/">Dark Mode</a></li>
