--- src/partials/home-header.html
@@ -1,39 +1,0 @@
-<div>
-  <nav class="navbar">
-    <!-- LOGO -->
-    <div class="logo">
-      <img
-        src="./images/booklogo.png"
-        rel="images"
-        max
-        width="70px"
-        alt="book"
-      />
-    </div>
-    <!-- NAVIGATION MENU -->
-    <ul class="nav-links">
-      <!-- USING CHECKBOX HACK -->
-      <input type="checkbox" id="checkbox_toggle" />
-      <label for="checkbox_toggle" class="hamburger">&#9776;</label>
-      <!-- NAVIGATION MENUS -->
-      <div class="menu">
-        <li>
-          <a href="">
-            <img src="./images/searchSmall.png" rel="images" alt="search" />
-          </a>
-        </li>
-        <li class="library">
-          <a href="/">My Library</a>
-          <!-- DROPDOWN MENU -->
-          <ul class="dropdown">
-            <li><a href="/">Books to Read </a></li>
-            <li><a href="/">Reading Now</a></li>
-            <li><a href="/">Finished Books</a></li>
-            <li><a href="library.html">My Library</a></li>
-          </ul>
-        </li>
-        <li><a href="/">Dark Mode</a></li>
-      </div>
-    </ul>
-  </nav>
-</div>
